requirements.txt
numpy>=1.19.2
pandas>=1.2.4
torch==2.6.0+cu126
torchvision==0.21.0+cu126scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
pillow>=8.2.0
opencv-python>=4.5.2
tqdm>=4.61.0
mlflow>=2.3.0
albumentations>=1.1.0
imbalanced-learn>=0.8.1
shap>=0.41.0